broker_gateway/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
httpx>=0.25.0
pydantic>=2.0.0
python-dotenv>=1.0.0

# 本地测试（使用 unittest + FastAPI TestClient，无需额外依赖）
